Restructuring (src/common/mixin => src/mixin)
