+++ src/mixin/index.js
@@ -1,4 +1,19 @@
 import style from './stylable';
+import definition from './definition';
+import fieldComponentBehaviour from './field-component-behaviour';
+import fieldGridBehaviour from './field-grid-behaviour';
+import referenceProperty from './reference-property';
+import storeBehaviour from './store-behaviour';
+import builtInComponents from './built-in-components';
+import ownIdentifier from './own-identifier';
+import i18n from './i18n/index';
+import { mixin as form } from './form/index';
+import { mixin as field } from './field/index';
 
-export { style };
-export default { style };
+export {
+    form, field, i18n, style, definition, fieldComponentBehaviour, fieldGridBehaviour, referenceProperty, storeBehaviour, builtInComponents, ownIdentifier
+};
+
+export default {
+    form, field, i18n, style, definition, fieldComponentBehaviour, fieldGridBehaviour, referenceProperty, storeBehaviour, builtInComponents, ownIdentifier
+};
