+++ src/style/index.scss
@@ -9,9 +9,9 @@
 @import '../application/loading-bar/style/loading-bar.scss';
 @import '../application/popin/style/popin.scss';
 
-//COMMON FOLDER
-@import '../common/field/style/field.scss';
-@import '../common/form/style/form.scss';
+//MIXIN FOLDER
+@import '../mixin/field/style/field.scss';
+@import '../mixin/form/style/form.scss';
 
 // COMPONENTS FOLDER
 @import '../components/button/style/button.scss';
